Feature/santiago agredo/add user repository test changes in user repository

+++ openmarket-server/src/main/java/co/unicauca/openmarket/server/domain/service/IOrderService.java
@@ -39,5 +39,5 @@ public interface IOrderService {
      * @param UserId id del usuario
      * @return lista de objetos entcontrados
      */
-    public List<Order> findByUser(Integer UserId);
+    public List<Order> findByUser(Long UserId);
 }
